+++ R/frames_graph.R
@@ -148,7 +148,7 @@ frames_graph <- function(m, r_list, r_times, r_type = "gradient", fade_raster = 
     ## create frames
     out("Creating frames...")
     if(graph_type == "flow"){
-      return(.gg_flow(m.split, gg.df, path_legend, path_legend_title, path_size, val_seq))
+      frames <- .gg_flow(m.split, gg.df, path_legend, path_legend_title, path_size, val_seq)
     }
     if(graph_type == "hist"){
       
@@ -177,8 +177,16 @@ frames_graph <- function(m, r_list, r_times, r_type = "gradient", fade_raster = 
       
       ## fusing histograms for plot scaling
       all.hist <- do.call(rbind, l.hist)
-      return(.gg_hist(l.hist, all.hist, path_legend, path_legend_title, path_size, val_seq, r_type))
+      frames <- .gg_hist(l.hist, all.hist, path_legend, path_legend_title, path_size, val_seq, r_type)
     }
   }
+  
+  ## add time attribute per frame
+  frames <- mapply(x = frames, y = unique(m.df$time), function(x, y){
+    attr(x, "time") <- y
+    return(x)
+  }, SIMPLIFY = F)
+  
+  return(frames)
 }